backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
tensorflow==2.15.0
numpy==1.22.3
Pillow==9.1.0
requests==2.27.1
python-dotenv==0.20.0
scikit-learn==1.0.2